Move file out of api wrapper directory since it is utility functions/constants, not strictly part of the wrapper
